Moved building off the ast into the top level
