+++ src/lex.rs
@@ -1,6 +1,7 @@
 //! Main module for lex library functionality
 
 pub mod ast;
+pub mod building;
 pub mod formats;
 pub mod lexing;
 pub mod parsing;
