+++ src/lex/parsing.rs
@@ -16,7 +16,6 @@
 
 // Parser implementations
 pub mod builder;
-pub mod building;
 pub mod common;
 pub mod ir;
 pub mod linebased;
